requirements.txt
python-telegram-bot>=21.7,<22
python-dotenv>=1.0.1,<2
sentence-transformers>=3.3.1,<4
ollama>=0.4.7,<1
